requirements.txt
paho-mqtt
pymodbus
watchdog
pyfiglet
pyModbusTCP
pyserial-asyncio
tomli